used BigDecimal instead of Double for cleaner Mapper

+++ src/main/kotlin/me/ezra_home/daily_snapshot/DailySnapshotUpdateDto.kt
@@ -1,17 +1,18 @@
 package me.ezra_home.daily_snapshot
 
+import java.math.BigDecimal
 import java.util.Optional
 import java.util.UUID
 
 data class DailySnapshotUpdateDto(
     val id: UUID? = null,
-    val startBalanceCash: Optional<Double>? = null,
-    val endBalanceCash: Optional<Double>? = null,
-    val outflowCash: Optional<Double>? = null,
-    val cogs: Optional<Double>? = null,
-    val cogsReturned: Optional<Double>? = null,
-    val expenses: Optional<Double>? = null,
-    val inflowJointAccount: Optional<Double>? = null,
-    val inflowPersonalAccount: Optional<Double>? = null,
-    val inflowCreditSales: Optional<Double>? = null
+    val startBalanceCash: Optional<BigDecimal>? = null,
+    val endBalanceCash: Optional<BigDecimal>? = null,
+    val outflowCash: Optional<BigDecimal>? = null,
+    val cogs: Optional<BigDecimal>? = null,
+    val cogsReturned: Optional<BigDecimal>? = null,
+    val expenses: Optional<BigDecimal>? = null,
+    val inflowJointAccount: Optional<BigDecimal>? = null,
+    val inflowPersonalAccount: Optional<BigDecimal>? = null,
+    val inflowCreditSales: Optional<BigDecimal>? = null
 )
